COMP: Avoid pandas incorrect usage future warnings
In future versions of pandas the access pattern used is
not going to be valid.  Replace with valid access and
replacement patterns.

```bash
python3 -Werror::FutureWarning -m pytest
```

```text
E   FutureWarning: A value is trying to be set on a copy of a DataFrame or Series through chained assignment using an inplace method.
E   The behavior will change in pandas 3.0. This inplace method will never work because the intermediate object on which we are setting values always behaves as a copy.
E
E   For example, when doing 'df[col].method(value, inplace=True)', try using 'df.method({col: value}, inplace=True)' or df[col] = df[col].method(value) instead, to perform the operation inplace on the original object.
```