BUG: Rare case of tracew with bvalue=0 from single series misclassified
Given a set of adjacent series (i.e. 405,406,407) which are tracew images
bug are classified as t2w, *and* one of them is a bvalue=0 image, identify that
an adjacent series image with bvalue>0 with identical physical space representation
exists, then rename the bvalue=0 image from t2w to tracew to put it in the family of
related images.